\name{abstract}
\docType{methods}
\alias{abstract}
\alias{abstract,Reference-method}

\title{Get abstracts from \code{Reference} object.}

\description{

Extract vector of abstracts of records in \code{Reference} object.

}

\section{Methods}{
\describe{

\item{\code{signature(object = "Reference")}}{
Vector of strings with name of elements corresponding to \code{id} of \code{Reference} object. If abstract is not available from record, then is \code{NA}.
 }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
