\name{Acronym}
\alias{Acronym}

\title{
Extracts \code{Acronym} from  \code{Medline} object.
}

\description{
Extractor for the \code{Acronym} of a \code{Medline} object.
}
\usage{
Acronym(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Acronym(myeloma)
}
