\name{Agency}
\alias{Agency}

\title{
Extracts \code{Agency} from  \code{Medline} object.
}

\description{
Extractor for the \code{Agency} of a \code{Medline} object.
}
\usage{
Agency(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Agency(myeloma)
}
