% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedlineC.R
\docType{methods}
\name{c,Medline-method}
\alias{c,ANY,ANY-method}
\alias{c,Medline-method}
\title{Concatenate Medline}
\usage{
\S4method{c}{Medline}(x, ...)
}
\arguments{
\item{x}{a Medline object}
\item{...}{Additional Medline objects to concatenate}
}
\value{
A Medline object
}
\description{
Concatenates multiple Medline objects
}

