\name{Citations}
\alias{Citations}

\title{
List of citing article titles where listed in PubMed.
}

\description{
Retrieves the number of citations in PubMed Central for the articles contained in the \code{Medline} object.
}

\usage{
Citations(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}
