\name{COIStatement}
\alias{COIStatement}

\title{
Extracts \code{COIStatement} from  \code{Medline} object.
}

\description{
Extractor for the \code{COIStatement} of a \code{Medline} object.
}
\usage{
COIStatement(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}
