\name{QueryCount}
\title{Count of record results for EUtilsQuery}
\usage{QueryCount(object)}
\description{Returns the number of Ids from a E-utility query}
\value{numeric count}
\seealso{\code{\link{EUtilsSummary}}}
\alias{QueryCount,EUtilsSummary-method}
\alias{QueryCount}
\author{Stephanie Kovalchik}
\arguments{\item{object}{object of the \code{EUtilsSummary} class.}}
