\name{Day}
\alias{Day}

\title{
Extracts \code{Day} from  \code{Medline} object.
}

\description{
Extractor for the \code{Day} of a \code{Medline} object.
}
\usage{
Day(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Day(myeloma)
}
