\name{Hour}
\alias{Hour}

\title{
Extracts \code{Hour} from  \code{Medline} object.
}

\description{
Extractor for the \code{Hour} of a \code{Medline} object.
}
\usage{
Hour(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Hour(myeloma)
}
