\name{QueryId}
\title{Get ids from EUtilsQuery}
\usage{QueryId(object)}
\description{Returns the vector Ids from a E-utility query}
\value{character vector of ids}
\seealso{\code{\link{EUtilsSummary}}}
\alias{QueryId,EUtilsSummary-method}
\alias{QueryId}
\author{Stephanie Kovalchik}
\arguments{\item{object}{object of the \code{EUtilsSummary} class.}}
