\name{keyword}
\docType{methods}
\alias{keyword}
\alias{keyword,Reference-method}

\title{Get keywords from \code{Reference} object.}

\description{

Extract list of keywords for records in \code{Reference} object.

}

\section{Methods}{
\describe{

\item{\code{signature(object = "Reference")}}{
List of vector of keywords with name of elements corresponding to \code{id} of \code{Reference} object. If not available from record, then entry is \code{NA}.
 }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
