\name{Language}
\alias{Language}

\title{
Extracts \code{Language} from  \code{Medline} object.
}

\description{
Extractor for the \code{Language} of a \code{Medline} object.
}
\usage{
Language(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Language(myeloma)
}
