\name{MedlineTA}
\alias{MedlineTA}

\title{
Extracts \code{MedlineTA} from  \code{Medline} object.
}

\description{
Extractor for the \code{MedlineTA} of a \code{Medline} object.
}
\usage{
MedlineTA(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
MedlineTA(myeloma)
}
