\name{Minute}
\alias{Minute}

\title{
Extracts \code{Minute} from  \code{Medline} object.
}

\description{
Extractor for the \code{Minute} of a \code{Medline} object.
}
\usage{
Minute(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Minute(myeloma)
}
