\name{Month}
\alias{Month}

\title{
Extracts \code{Month} from  \code{Medline} object.
}

\description{
Extractor for the \code{Month} of a \code{Medline} object.
}
\usage{
Month(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Month(myeloma)
}
