\name{myeloma}
\docType{data}
\alias{myeloma}
\title{Example of results from \code{EUtilsGet}}

\description{
First five records of query of NCBI PubMed for "myeloma[MeSh]" in the past year (at the time of this writing) as a Medline object.
}

\usage{myeloma}

\format{A list object with five PubMed article records.}

\source{PubMed database of the US National Library of Medicine}

\seealso{\code{\link{EUtilsGet}},\code{\link{Medline}}}

\author{Stephanie Kovalchik \email{s.a.kovalchik@gmail.com}}
