\name{print-Medline}
\docType{methods}
\alias{print,Medline-method}

\title{ Print method for Medline object }

\description{

Print object of class \code{Medline}

}

\section{Methods}{

\describe{

\item{\code{signature(object = "Medline")}}{
Query translation (if provided) and total fetched record count for \code{Medline} object.
}

}
}

\keyword{methods}

