\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,EUtilsSummary-method}

\title{ Print method for EUtilsSummary object }

\description{

Print object of class \code{EUtilsSummary}

}

\section{Methods}{

\describe{

\item{\code{signature(object = "EUtilsSummary")}}{
Query translation for \code{EUtilsSummary} object.
}

}
}
\keyword{methods}

