\name{RISmed-package}
\alias{RISmed-package}
\alias{RISmed}

\docType{package}

\title{

Download content from NCBI databases.


}

\description{
 Download content from NCBI databases. This package is intended to facilitate analyses of NCBI database content. This package is not intended for reference management.
}

\details{
\tabular{ll}{
Package: \tab RISmed\cr
Type: \tab Package\cr
Version: \tab 2.1.3\cr
Date: \tab 2014-7-20\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

Note on usage:

In order not to overload the E-utility servers, NCBI recommends that users post no more than three URL requests per second and limit large jobs to either weekends or between 9:00 PM and 5:00 AM Eastern time during weekdays. Failure to comply with this policy may result in an IP address being blocked from accessing NCBI. 

}
\author{

Maintainer: Stephanie Kovalchik <s.a.kovalchik@gmail.com>

}
