\name{show-Medline}
\docType{methods}
\alias{show,Medline-method}

\title{ Show method for Medline object }

\description{

Show object of class \code{Medline}

}

\section{Methods}{

\describe{

\item{\code{signature(object = "Medline")}}{
Query translation (if provided) and total fetched record count for \code{Medline} object.
}

}
}
\keyword{methods}

