\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,Reference-method}

\title{ Methods for Function show in Package `RISmed' }

\description{

Show object of class \code{Reference}

}

\section{Methods}{

\describe{

\item{\code{signature(object = "Reference")}}{
Abbreviated data.frame of \code{Reference} object.
}

}
}
\keyword{methods}

