\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,Reference-method}

\title{ Methods for Function summary in Package `RISmed' }

\description{

Summary of object of class \code{Reference}

}

\section{Methods}{
\describe{

\item{\code{signature(object = "Reference",...)}}{
Returns the number of records for \code{Reference} object and frequency counts of records by year.
  }
 }
}

\keyword{methods}

