\name{Title}
\alias{Title}

\title{
Extracts \code{Title} from  \code{Medline} object.
}

\description{
Extractor for the \code{Title} of a \code{Medline} object.
}
\usage{
Title(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Title(myeloma)
}
