\name{translation}
\title{Get the NCBI query translation from EUtilsQuery}
\usage{translation(object)}
\description{Returns the string translation of a E-utility query}
\value{character query translation}
\seealso{\code{\link{EUtilsSummary}}}
\alias{translation,EUtilsSummary-method}
\alias{translation}
\author{Stephanie Kovalchik}
\arguments{\item{object}{object of the \code{EUtilsSummary} class.}}
