\name{Volume}
\alias{Volume}

\title{
Extracts \code{Volume} from  \code{Medline} object.
}

\description{
Extractor for the \code{Volume} of a \code{Medline} object.
}
\usage{
Volume(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Volume(myeloma)
}
