\name{Year}
\alias{Year}

\title{
Extracts \code{Year} from  \code{Medline} object.
}

\description{
Extractor for the \code{Year} of a \code{Medline} object.
}
\usage{
Year(object)
}

\arguments{
  \item{object}{instance of class \code{Medline}}
}

\value{vector}

\seealso{\code{\link{Medline}}}

\examples{
data(myeloma)
Year(myeloma)
}
