\name{count}
\title{Count of record results for EUtilsQuery}
\usage{count(object)}
\description{Returns the number of Ids from a E-utility query}
\value{numeric count}
\seealso{\code{\link{EUtilsSummary}}}
\alias{count,EUtilsSummary-method}
\alias{count}
\author{Stephanie Kovalchik}
\arguments{\item{object}{object of the \code{EUtilsSummary} class.}}
