\name{id}
\title{Get ids from EUtilsQuery}
\usage{id(object)}
\description{Returns the vector Ids from a E-utility query}
\value{character vector of ids}
\seealso{\code{\link{EUtilsSummary}}}
\alias{id,EUtilsSummary-method}
\alias{id}
\author{Stephanie Kovalchik}
\arguments{\item{object}{object of the \code{EUtilsSummary} class.}}
