\name{RISmed-package}
\alias{RISmed-package}
\alias{RISmed}

\docType{package}

\title{

Download content from NCBI databases.


}

\description{
 Download content from NCBI databases. This package is intended to facilitate analyses of NCBI database content. This package is not intended for reference management.
}

\details{
\tabular{ll}{
Package: \tab RISmed\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2011-03-23\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{

Maintainer: Stephanie Kovalchik <s.a.kovalchik@gmail.com>

}
