\name{RISmed-package}
\alias{RISmed-package}
\alias{RISmed}

\docType{package}

\title{

Search PubMed; Make bibliographic content into analyzable objects

}

\description{

Perform PubMed search, create \code{Reference} object from search results or RIS file.

}

\details{
\tabular{ll}{
Package: \tab RISmed\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2011-03-23\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{

Maintainer: S. Kovalchik <s.a.kovalchik@gmail.com>

}
