\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,EUtilsSummary-method}

\title{ Show method for EUtilsSummary object }

\description{

Show object of class \code{EUtilsSummary}

}

\section{Methods}{

\describe{

\item{\code{signature(object = "EUtilsSummary")}}{
Query translation for \code{EUtilsSummary} object.
}

}
}
\keyword{methods}

