\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,EUtilsSummary-method}

\title{ Summary method for EUtilsSummary object}

\description{

Summary of object of class \code{EUtilsSummary}

}

\section{Methods}{
\describe{

\item{\code{signature(object = "EUtilsSummary",...)}}{
Returns the number of records and id vector.
  }
 }
}

\keyword{methods}

