\name{QueryTranslation}
\title{Get the NCBI query translation from EUtilsQuery}
\usage{QueryTranslation(object)}
\description{Returns the string translation of a E-utility query}
\value{character query translation}
\seealso{\code{\link{EUtilsSummary}}}
\alias{QueryTranslation,EUtilsSummary-method}
\alias{QueryTranslation}
\author{Stephanie Kovalchik}
\arguments{\item{object}{object of the \code{EUtilsSummary} class.}}
